Implemented the collidesWith method within BaseActor. Removed xCoord and yCoord field redundancies. Trickled the "double" data type change from BaseActor up to Actor. Commented out setUp sprite temporarily.

+++ src/com/edd/circlebrawl/Actor.java
@@ -1,26 +1,32 @@
 package com.edd.circlebrawl;
+
 /*
  * This is going to be the interface for the base Actor.
  * Interfaces can't have non-constant or non static member
  * variables so I'm going to just use this interface to 
  * guarantee some methods that every that implements 
  * Actor is going to be required to have.*/
-import acm.graphics.*;
+import acm.graphics.GImage;
 
-public interface Actor 
-{
-	//getters
-	public int getX();
-	public int getY();
+public interface Actor {
+	// getters
+	public double getX();
+
+	public double getY();
+
 	public GImage getSprite();
-	
-	//setters
+
+	// setters
 	public void setX(int x);
+
 	public void setY(int y);
+
 	public void setSprite(GImage sprite);
-	
-	//other
+
+	// other
 	public boolean collidesWith(BaseActor anotherActor);
+
 	public void remove();
+
 	public void setupSprite(String spriteFile);
 }
